feat(order): Implement order processing and cash management
- Implements OrderService to handle the creation and management of orders.

- Adds related repositories: OrderRepository and OrderItemRepository.

- Introduces new DTOs for order creation and cash charging: OrderRequest, OrderItemRequest, and ChargeCashRequest.

- Updates domain models (Order, OrderItem, User) to support the order process and cash transactions.

- Implements commission policy logic for future use.

- Refactors OrderStatus to a new lifecycle.

+++ src/main/java/mateandgit/opener_maket/dto/AddItemRequest.java
@@ -2,12 +2,14 @@ package mateandgit.opener_maket.dto;
 
 import mateandgit.opener_maket.domain.status.CategoryStatus;
 
+import java.math.BigDecimal;
+
 public record AddItemRequest(
         String email,
         CategoryStatus category,
         String itemName,
         String description,
-        int price,
+        BigDecimal price,
         int stockQuantity
 
 ) {
